chore(firmware-authenticity): extract type utils & guards
