+++ suite-common/firmware-authenticity/src/index.ts
@@ -1,1 +1,2 @@
 export * from './scenariosConfig';
+export * from './utils';
