+++ suite-common/firmware-authenticity/src/scenariosConfig.ts
@@ -4,7 +4,6 @@ import {
     FirmwareRevisionCheckError,
 } from '@trezor/connect';
 import { DeviceModelInternal } from '@trezor/device-utils';
-import type { FilterPropertiesByType } from '@trezor/type-utils';
 
 /*
  * Various scenarios how firmware authenticity check errors are handled in Suite and Suite Lite
@@ -97,25 +96,6 @@ export const hashCheckErrorScenarios = {
     },
 } satisfies HashCheckErrorScenarios;
 
-// TODO: when this is moved to suite-common, extract the following derived types & typeguards to a new file
-export type SkippedRevisionCheckError = keyof FilterPropertiesByType<
-    typeof revisionCheckErrorScenarios,
-    { type: 'skipped' }
->;
-
-export const isSkippedRevisionCheckError = (
-    error: FirmwareRevisionCheckError,
-): error is SkippedRevisionCheckError => revisionCheckErrorScenarios[error].type === 'skipped';
-
-export type SkippedHashCheckError = keyof FilterPropertiesByType<
-    typeof hashCheckErrorScenarios,
-    { type: 'skipped' }
->;
-
-export const isSkippedHashCheckError = (
-    error: FirmwareHashCheckError,
-): error is SkippedHashCheckError => hashCheckErrorScenarios[error].type === 'skipped';
-
 // stricter type than required by @trezor/connect, which is very permissive, being a library
 type ExhaustiveFirmwareHashCheckTimeouts = FirmwareHashCheckTimeouts &
     Record<DeviceModelInternal, number | undefined>;
@@ -129,24 +109,3 @@ export const FW_HASH_CHECK_DEFAULT_TIMEOUTS: ExhaustiveFirmwareHashCheckTimeouts
     [DeviceModelInternal.T3W1]: undefined,
     [DeviceModelInternal.UNKNOWN]: undefined,
 };
-
-export type RevisionCheckErrorWithNotification = keyof FilterPropertiesByType<
-    typeof revisionCheckErrorScenarios,
-    { shouldNotify: true }
->;
-
-export const isRevisionCheckErrorWithNotification = (
-    error: FirmwareRevisionCheckError,
-): error is RevisionCheckErrorWithNotification =>
-    revisionCheckErrorScenarios[error].shouldNotify === true;
-
-export type HashCheckErrorWithNotification = keyof FilterPropertiesByType<
-    typeof hashCheckErrorScenarios,
-    { shouldNotify: true }
->;
-
-export const isHashCheckErrorWithNotification = (
-    error: FirmwareHashCheckError,
-): error is HashCheckErrorWithNotification =>
-    // @ts-expect-error if this no longer gives error, then TODO hash check notifications must be implemented
-    hashCheckErrorScenarios[error].shouldNotify === true;
